Added AuthService and AuthProvider to handle the web authentication flow on it's own, decoupled from Login Service. Updated OAuthCodeTask as per PR comments.
